Support test-all when finagle-clojure is used
The test-all target chooses the project.clj file of the first internal
dep as a base file, and then composes all the other project.clj files
considering certain keys in the merge process. If you are unluckly
enough to have the first internal dep be a project that specifies
:finagle-clojure, then that plugin will activate before the tests
are run, since you are likely not in the base project when running
test-all, nor are you likely in a project that itself uses
:finagle-clojure, you end up with no thrift files to compile (which
you didn't want to occur anyway), and things derail.

The fix is to just dissoc this key from the base project.